requirements.txt
deta==1.1.0
fastapi==0.92.0
pydantic==1.10.4
python-dotenv==1.0.0
requests==2.22.0
starlette==0.25.0
